+++ KmpSqlencrypt/src/androidAndroidTest/kotlin/com/oldguy/kiscmp/BasicTests.kt
@@ -1,76 +1,62 @@
 package com.oldguy.kiscmp
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import androidx.test.platform.app.InstrumentationRegistry
 import com.oldguy.database.SqlValue
 import org.junit.Test
 import org.junit.runner.RunWith
 import kotlinx.coroutines.*
+import kotlinx.coroutines.test.runTest
 import kotlin.test.assertEquals
 
+@ExperimentalCoroutinesApi
 @RunWith(AndroidJUnit4::class)
 class AndroidJunitTests: SqlCipherTests() {
 
+    /**
+     * Run all unencrypted basic tests on UTF-8 encoded in-memory DB
+     */
+    @ExperimentalStdlibApi
     @Test
-    fun openCloseTest() {
+    fun testOpenClose() {
         db = createDb
         if (mapBooleanYN)
             SqlValue.BooleanValue.mapping("N", "Y")
-        runBlocking {
+        runTest {
             db.use("") {
-                testVersions()
-                pragmaTest()
-                script1Test()
-                verifyTest1Table()
-
-                testBindInsert()
-                testSelect()
-
-                testUpdate()
-                testUnnamedInsert()
-                testDelete()
-
-                testTableTest2()
-                testTimestamps()
+                allTests()
             }
         }
     }
 
+    /**
+     * Run all unencrypted basic tests on UTF-16 big endian encoded in-memory DB
+     */
     @Test
-    fun encryption1Test() {
-        val dbName = "KeyTest1.db"
-        val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        appContext.deleteDatabase(dbName)
-        val path = appContext.getDatabasePath(dbName).absolutePath
-        runBlocking {
-            testPasswordsAndUpgrade(path)
+    fun testOpenCloseUtf16be() {
+        db = createDb16BE
+        if (mapBooleanYN)
+            SqlValue.BooleanValue.mapping("N", "Y")
+        runTest {
+            db.use("") {
+                assertEquals(SqliteEncoding.Utf16BigEndian, db.queryEncoding())
+                allTests()
+            }
         }
     }
 
+    /**
+     * Run all unencrypted basic tests on UTF-16 little endian encoded in-memory DB
+     */
     @Test
-    fun openCloseUtf16leTest() {
+    fun testOpenCloseUtf16le() {
         db = createDb16LE
         if (mapBooleanYN)
             SqlValue.BooleanValue.mapping("N", "Y")
-        runBlocking {
+        runTest {
             db.use("") {
-                testVersions()
-                assertEquals("UTF-16le", db.encoding.pragma)
-                pragmaTest()
-                script1Test()
-                verifyTest1Table()
-
-                testBindInsert()
-                testSelect()
-
-                testUpdate()
-                testUnnamedInsert()
-                testDelete()
-
-                testTableTest2()
-                testTimestamps()
+                assertEquals(SqliteEncoding.Utf16LittleEndian, db.queryEncoding())
+                allTests()
             }
         }
     }
-
 }