+++ Changelog.md
@@ -1,5 +1,6 @@
 ## KmpSqlencrypt Change Log
 
+
 ** 0.4.5 ** 2022-06
 
 No API changes, only release upgrades.
@@ -13,6 +14,12 @@ No API changes, only release upgrades.
 - kotlinx coroutines 1.6.2
 - Android NDK 25.0.8528842
 - Android build tools 33.0.0
+- Gradle 7.4.2
+
+**0.4.4**  2022-03
+
+- SqlCipher 4.5.1 (Sqlite 3.37.2) 
+- Correct native code supporting strings when database encoding is configured for a 16 bit charset.
 
 **0.4.3**  2022-02-02